requirements/base.txt
# Core Django packages
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
mysqlclient==2.2.4
django-environ==0.11.2

# Message Queue (RabbitMQ)
pika==1.3.2

# Cache (Redis)
redis==5.0.3
django-redis==5.4.0



# CORS
django-cors-headers==4.3.1

# Monitoring & Metrics
django-prometheus==2.3.1

# API Documentation
drf-spectacular==0.27.2

# HTTP Client for external APIs
requests==2.31.0
httpx==0.27.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Security
cryptography==42.0.5

# Validation
jsonschema==4.21.1

# Production WSGI server
gunicorn==21.2.0

# Health checks
django-health-check==3.18.1
